requirements.txt
fbchat==1.9.7
openai==0.27.0
Flask==2.0.3
requests==2.26.0
python-dotenv==0.19.2